improve asciidocs
