+++ controllers/flp/flp_pipeline_builder.go
@@ -17,6 +17,7 @@ import (
 	"github.com/netobserv/network-observability-operator/controllers/flp/fmstatus"
 	"github.com/netobserv/network-observability-operator/pkg/conversion"
 	"github.com/netobserv/network-observability-operator/pkg/helper"
+	otelConfig "github.com/netobserv/network-observability-operator/pkg/helper/otel"
 	"github.com/netobserv/network-observability-operator/pkg/loki"
 	"github.com/netobserv/network-observability-operator/pkg/metrics"
 	"github.com/netobserv/network-observability-operator/pkg/volumes"
@@ -254,8 +255,8 @@ func (b *PipelineBuilder) AddProcessorStages() error {
 		enrichedStage.EncodePrometheus("prometheus", promEncode)
 	}
 
-	b.addCustomExportStages(&enrichedStage, flpMetrics)
-	return nil
+	err := b.addCustomExportStages(&enrichedStage, flpMetrics)
+	return err
 }
 
 func flowMetricToFLP(flowMetric *metricslatest.FlowMetricSpec) (*api.MetricsItem, error) {
@@ -473,7 +474,7 @@ func (b *PipelineBuilder) addTransformFilter(lastStage config.PipelineBuilderSta
 	return lastStage
 }
 
-func (b *PipelineBuilder) addCustomExportStages(enrichedStage *config.PipelineBuilderStage, flpMetrics []api.MetricsItem) {
+func (b *PipelineBuilder) addCustomExportStages(enrichedStage *config.PipelineBuilderStage, flpMetrics []api.MetricsItem) error {
 	for i, exporter := range b.desired.Exporters {
 		if exporter.Type == flowslatest.KafkaExporter {
 			b.createKafkaWriteStage(fmt.Sprintf("kafka-export-%d", i), &exporter.Kafka, enrichedStage)
@@ -482,9 +483,13 @@ func (b *PipelineBuilder) addCustomExportStages(enrichedStage *config.PipelineBu
 			createIPFIXWriteStage(fmt.Sprintf("IPFIX-export-%d", i), &exporter.IPFIX, enrichedStage)
 		}
 		if exporter.Type == flowslatest.OpenTelemetryExporter {
-			b.createOpenTelemetryStage(fmt.Sprintf("Otel-export-%d", i), &exporter.OpenTelemetry, enrichedStage, flpMetrics)
+			err := b.createOpenTelemetryStage(fmt.Sprintf("Otel-export-%d", i), &exporter.OpenTelemetry, enrichedStage, flpMetrics)
+			if err != nil {
+				return err
+			}
 		}
 	}
+	return nil
 }
 
 func (b *PipelineBuilder) createKafkaWriteStage(name string, spec *flowslatest.FlowCollectorKafka, fromStage *config.PipelineBuilderStage) config.PipelineBuilderStage {
@@ -518,7 +523,7 @@ func getIPFIXTransport(transport string) string {
 	}
 }
 
-func (b *PipelineBuilder) createOpenTelemetryStage(name string, spec *flowslatest.FlowCollectorOpenTelemetry, fromStage *config.PipelineBuilderStage, flpMetrics []api.MetricsItem) {
+func (b *PipelineBuilder) createOpenTelemetryStage(name string, spec *flowslatest.FlowCollectorOpenTelemetry, fromStage *config.PipelineBuilderStage, flpMetrics []api.MetricsItem) error {
 	conn := api.OtlpConnectionInfo{
 		Address:        spec.TargetHost,
 		Port:           spec.TargetPort,
@@ -532,7 +537,11 @@ func (b *PipelineBuilder) createOpenTelemetryStage(name string, spec *flowslates
 
 	if logsEnabled || metricsEnabled {
 		// add transform stage
-		transformStage := fromStage.TransformGeneric(fmt.Sprintf("%s-transform", name), helper.GetOtelTransformConfig(spec.FieldsMapping))
+		transformCfg, err := otelConfig.GetOtelTransformConfig(spec.FieldsMapping)
+		if err != nil {
+			return err
+		}
+		transformStage := fromStage.TransformGeneric(fmt.Sprintf("%s-transform", name), *transformCfg)
 
 		// otel logs config
 		if logsEnabled {
@@ -544,10 +553,14 @@ func (b *PipelineBuilder) createOpenTelemetryStage(name string, spec *flowslates
 
 		// otel metrics config
 		if metricsEnabled {
+			metricsCfg, err := otelConfig.GetOtelMetrics(flpMetrics)
+			if err != nil {
+				return err
+			}
 			transformStage.EncodeOtelMetrics(fmt.Sprintf("%s-metrics", name), api.EncodeOtlpMetrics{
 				OtlpConnectionInfo: &conn,
 				Prefix:             "netobserv_",
-				Metrics:            helper.GetOtelMetrics(flpMetrics),
+				Metrics:            metricsCfg,
 				PushTimeInterval:   api.Duration{Duration: spec.Metrics.PushTimeInterval.Duration},
 				ExpiryTime:         api.Duration{Duration: 2 * time.Minute},
 			})
@@ -555,6 +568,7 @@ func (b *PipelineBuilder) createOpenTelemetryStage(name string, spec *flowslates
 
 		// TODO: implement api.EncodeOtlpTraces
 	}
+	return nil
 }
 
 func getOtelConnType(connType string) string {
