+++ hack/asciidoc-flows-gen.sh
@@ -1,27 +1,31 @@
 #!/bin/sh
 
 ADOC=docs/flows-format.adoc
-SOURCE=controllers/consoleplugin/config/static-frontend-config.yaml
+FE_SOURCE=controllers/consoleplugin/config/static-frontend-config.yaml
+CARDINALITY_SOURCE=pkg/helper/cardinality/cardinality.json
+OTEL_SOURCE=pkg/helper/otel/otel-config.json
 
 # Header
 echo "// Automatically generated by '$0'. Do not edit, or make the NETOBSERV team aware of the editions." > $ADOC
 cat ./hack/flows-format-header.adoc >> $ADOC
 echo -e "\n" >> $ADOC
 
-echo -e '[cols="1,1,3,1,1,1",options="header"]' >> $ADOC
+echo -e '[cols="1,1,3,1,1,1,1",options="header"]' >> $ADOC
 echo -e '|===' >> $ADOC
-echo -e '| Name | Type | Description | Filter ID | Loki label | Cardinality' >> $ADOC
+echo -e '| Name | Type | Description | Filter ID | Loki label | Cardinality | Otel' >> $ADOC
 
-nbfields=$(yq '.fields | length' $SOURCE)
+nbfields=$(yq '.fields | length' $FE_SOURCE)
+cardinalityMap=$(cat $CARDINALITY_SOURCE)
+otelMap=$(cat $OTEL_SOURCE)
 
 for i in $(seq 0 $(( $nbfields-1 )) ); do
-  entry=$(yq ".fields | sort_by(.name) | .[$i]" $SOURCE)
-  name=$(printf "$entry" | yq ".name")
-  type=$(printf "$entry" | yq ".type")
-  desc=$(printf "$entry" | yq ".description")
-  filter=$(printf "$entry" | yq ".filter")
+  frontEntry=$(yq ".fields | sort_by(.name) | .[$i]" $FE_SOURCE)
+  name=$(printf "$frontEntry" | yq ".name")
+  type=$(printf "$frontEntry" | yq ".type")
+  desc=$(printf "$frontEntry" | yq ".description")
+  filter=$(printf "$frontEntry" | yq ".filter")
   if [[ "$filter" == "null" ]]; then
-    filter=$(yq ".columns[] | select(.field == \"$name\").filter" $SOURCE | sed 's/null//')
+    filter=$(yq ".columns[] | select(.field == \"$name\").filter" $FE_SOURCE | sed 's/null//')
     if [[ "$filter" == "" ]]; then
       filter="n/a"
     else
@@ -30,19 +34,28 @@ for i in $(seq 0 $(( $nbfields-1 )) ); do
   else 
     filter="\`$filter\`"
   fi
-  isLabel=$(printf "$entry" | yq ".lokiLabel")
+  isLabel=$(printf "$frontEntry" | yq ".lokiLabel")
   if [[ $isLabel == "true" ]]; then
     isLabel="yes"
   else
     isLabel="no"
   fi
-  cardWarn=$(printf "$entry" | yq ".cardinalityWarn")
+  cardWarn=$(printf "$cardinalityMap" | jq -r ".$name")
+  if [[ "$cardWarn" == "null" ]]; then
+      echo "missing cardinality for field $name"
+      exit 1
+  fi
+  otel=$(printf "$otelMap" | jq -r ".$name")
+  if [[ "$otel" == "null" ]]; then
+      otel="n/a"
+  fi
   echo -e "| \`$name\`" >> $ADOC
   echo -e "| $type" >> $ADOC
   echo -e "| $desc" >> $ADOC
   echo -e "| $filter" >> $ADOC
   echo -e "| $isLabel" >> $ADOC
   echo -e "| $cardWarn" >> $ADOC
+  echo -e "| $otel" >> $ADOC
 done
 
 echo -e '|===' >> $ADOC
