+++ pkg/helper/cardinality/cardinality.json
@@ -1,0 +1,56 @@
+{
+  "TimeFlowStartMs": "avoid",
+  "TimeFlowEndMs": "avoid",
+  "TimeReceived": "avoid",
+  "SrcK8S_Name": "careful",
+  "SrcK8S_Type": "fine",
+  "SrcK8S_OwnerName": "fine",
+  "SrcK8S_OwnerType": "fine",
+  "SrcK8S_Namespace": "fine",
+  "SrcAddr": "avoid",
+  "SrcPort": "careful",
+  "SrcMac": "avoid",
+  "SrcK8S_HostIP": "fine",
+  "SrcK8S_HostName": "fine",
+  "SrcK8S_Zone": "fine",
+  "SrcSubnetLabel": "fine",
+  "DstK8S_Name": "careful",
+  "DstK8S_Type": "fine",
+  "DstK8S_OwnerName": "fine",
+  "DstK8S_OwnerType": "fine",
+  "DstK8S_Namespace": "fine",
+  "DstAddr": "avoid",
+  "DstPort": "careful",
+  "DstMac": "avoid",
+  "DstK8S_HostIP": "fine",
+  "DstK8S_HostName": "fine",
+  "DstK8S_Zone": "fine",
+  "DstSubnetLabel": "fine",
+  "K8S_FlowLayer": "fine",
+  "Proto": "fine",
+  "Dscp": "fine",
+  "IcmpType": "fine",
+  "IcmpCode": "fine",
+  "Duplicate": "fine",
+  "FlowDirection": "fine",
+  "IfDirections": "fine",
+  "Interfaces": "careful",
+  "Flags": "fine",
+  "Bytes": "avoid",
+  "Packets": "avoid",
+  "PktDropBytes": "avoid",
+  "PktDropPackets": "avoid",
+  "PktDropLatestState": "fine",
+  "PktDropLatestDropCause": "fine",
+  "PktDropLatestFlags": "fine",
+  "DnsId": "avoid",
+  "DnsLatencyMs": "avoid",
+  "DnsFlags": "fine",
+  "DnsFlagsResponseCode": "fine",
+  "DnsErrno": "fine",
+  "TimeFlowRttNs": "avoid",
+  "NetworkEvents": "avoid",
+  "K8S_ClusterName": "fine",
+  "_RecordType": "fine",
+  "_HashId": "avoid"
+}
