+++ pkg/helper/otel/otel-config.json
@@ -1,0 +1,50 @@
+{
+  "SrcAddr": "source.address",
+  "SrcMac": "source.mac",
+  "SrcPort": "source.port",
+  "SrcK8S_Name": "source.k8s.name",
+  "SrcK8S_Type": "source.k8s.kind",
+  "SrcK8S_OwnerName": "source.k8s.owner.name",
+  "SrcK8S_OwnerType": "source.k8s.owner.kind",
+  "SrcK8S_Namespace": "source.k8s.namespace.name",
+  "SrcK8S_HostIP": "source.k8s.host.address",
+  "SrcK8S_HostName": "source.k8s.host.name",
+  "SrcK8S_Zone": "source.zone",
+  "DstAddr": "destination.address",
+  "DstMac": "destination.mac",
+  "DstPort": "destination.port",
+  "DstK8S_Name": "destination.k8s.name",
+  "DstK8S_Type": "destination.k8s.kind",
+  "DstK8S_OwnerName": "destination.k8s.owner.name",
+  "DstK8S_OwnerType": "destination.k8s.owner.kind",
+  "DstK8S_Namespace": "destination.k8s.namespace.name",
+  "DstK8S_HostIP": "destination.k8s.host.address",
+  "DstK8S_HostName": "destination.k8s.host.name",
+  "DstK8S_Zone": "destination.zone",
+  "Bytes": "bytes",
+  "Packets": "packets",
+  "Proto": "protocol",
+  "Flags": "tcp.flags",
+  "TimeFlowRttNs": "tcp.rtt",
+  "Interfaces": "interface.names",
+  "IfDirections": "interface.directions",
+  "FlowDirection": "host.direction",
+  "DnsErrno": "dns.errno",
+  "DnsFlags": "dns.flags",
+  "DnsFlagsResponseCode": "dns.responsecode",
+  "DnsId": "dns.id",
+  "DnsLatencyMs": "dns.latency",
+  "Dscp": "dscp",
+  "IcmpCode": "icmp.code",
+  "IcmpType": "icmp.type",
+  "K8S_ClusterName": "k8s.cluster.name",
+  "K8S_FlowLayer": "k8s.layer",
+  "PktDropBytes": "drops.bytes",
+  "PktDropPackets": "drops.packets",
+  "PktDropLatestDropCause": "drops.latestcause",
+  "PktDropLatestFlags": "drops.latestflags",
+  "PktDropLatestState": "drops.lateststate",
+  "TimeFlowEndMs": "timeflowend",
+  "TimeFlowStartMs": "timeflowstart",
+  "TimeReceived": "timereceived"
+}
